Remove autocmd for CursorMoved[I] to ease navigation
Original code for using autocmd for CursorMoved[I] causes interface to
stall often especially when user navigates by pressing and holding
j key(I know it is a bad practice). But highlighting of parenthesis was
fast!.

Removing CursorMoved[I] autocmds overcomes this issue. On the downside,
updating of screen now depends on updatetime global setting. We can live
with this for now.

Signed-off-by: Ratheesh S <ratheeshreddy@gmail.com>